refactor: dto, outcome 패키지 생성, OutcomeStatics를 OutcomeSummaryDto로 변경
